fix: seed the row tab appends to a table
